profile and privacy

+++ app/Http/Controllers/UserOrderController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Gate;
 use Auth;
 use App\UserOrder;
 use Illuminate\Http\Request;
@@ -9,10 +10,16 @@ use Illuminate\Http\Request;
 class UserOrderController extends Controller
 {
     public function showUserOrderForm(){
+        if(Gate::allows('isShop'))
+            return back();
+
         return view('profile.user_order_form');
     }
 
     public function create(Request $request){
+        if(Gate::allows('isShop'))
+            return back();
+        
         $this->validate($request, [
             'spare_part_id' => 'required|exists:spare_parts,id',
             'min_price' => 'required|integer|min:0',
@@ -29,7 +36,17 @@ class UserOrderController extends Controller
         return redirect()->route('profile');
     }
 
+    public function get(UserOrder $order){
+        if(Gate::denies('isOwner', $order))
+            return back();
+
+        return view('profile.user_order', ['order' => $order]);
+    }
+
     public function delete(UserOrder $order){
+        if(Gate::denies('isOwner', $order))
+            return back();
+
         $order->delete();
 
         return redirect()->route('profile');
